+++ packages/qw-page/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.2.1] - 25/03/2021
+
+### Updated
+
+- dependencies
+- webpack config
+- exported filename
+- package.json
+
 ## [0.2.0] - 25/03/2021
 
 ### Added